requirements.txt
Django==4.2.4
djangorestframework
djangorestframework-simplejwt
mysqlclient